refactor(remove-usecases): refactor domain layer
