+++ lib/application/workout/workout_bloc.dart
@@ -8,16 +8,16 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:power_progress/application/exercise/month/month_bloc.dart';
 import 'package:power_progress/application/exercise/week/week_bloc.dart';
 import 'package:power_progress/application/one_rm/one_rm_bloc.dart' as or_bloc;
-import 'package:power_progress/domain/core/entities/value_objects/month.dart';
-import 'package:power_progress/domain/core/entities/value_objects/one_rm.dart';
-import 'package:power_progress/domain/core/entities/week_enum.dart';
+import 'package:power_progress/domain/core/value_objects/month.dart';
+import 'package:power_progress/domain/core/value_objects/one_rm.dart';
+import 'package:power_progress/domain/core/week_enum.dart';
 import 'package:power_progress/domain/exercise/value_objects/incrementation.dart';
 import 'package:power_progress/domain/one_rm/one_rm_failure.dart';
 import 'package:power_progress/domain/one_rm/i_one_rm_repository.dart';
-import 'package:power_progress/domain/workout/entities/month_workout.dart';
-import 'package:power_progress/domain/workout/entities/workout.dart';
-import 'package:power_progress/domain/workout/entities/workout_failure.dart';
-import 'package:power_progress/domain/workout/repositories/i_workout_repository.dart';
+import 'package:power_progress/domain/workout/month_workout.dart';
+import 'package:power_progress/domain/workout/workout.dart';
+import 'package:power_progress/domain/workout/workout_failure.dart';
+import 'package:power_progress/domain/workout/i_workout_repository.dart';
 
 part 'workout_event.dart';
 part 'workout_state.dart';
