+++ lib/domain/core/week_enum.dart
