+++ lib/domain/core/week_enum.freezed.dart
