+++ lib/domain/exercise/value_objects/week.dart
@@ -2,7 +2,7 @@ import 'package:dartz/dartz.dart';
 
 import 'package:power_progress/core/domain/value_failure.dart';
 import 'package:power_progress/core/domain/value_object.dart';
-import 'package:power_progress/domain/core/entities/week_enum.dart';
+import 'package:power_progress/domain/core/week_enum.dart';
 
 class Week extends ValueObject<WeekEnum> {
   @override
