+++ lib/domain/workout/amrap_exercise_set.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 
-import 'package:power_progress/domain/workout/entities/exercise_set.dart';
-import 'package:power_progress/domain/workout/entities/value_objects/weight.dart';
+import 'package:power_progress/domain/workout/exercise_set.dart';
+import 'package:power_progress/domain/workout/value_objects/weight.dart';
 
 class AmrapExerciseSet extends ExerciseSet {
   final bool isDone;
