+++ lib/domain/workout/i_workout_repository.dart
@@ -1,9 +1,9 @@
 import 'package:dartz/dartz.dart';
-import 'package:power_progress/domain/core/entities/value_objects/month.dart';
 
-import 'package:power_progress/domain/core/entities/week_enum.dart';
-import 'package:power_progress/domain/workout/entities/workout_done.dart';
-import 'package:power_progress/domain/workout/entities/workout_failure.dart';
+import 'package:power_progress/domain/core/value_objects/month.dart';
+import 'package:power_progress/domain/core/week_enum.dart';
+import 'package:power_progress/domain/workout/workout_done.dart';
+import 'package:power_progress/domain/workout/workout_failure.dart';
 
 abstract class IWorkoutRepository {
   Future<Either<WorkoutFailure, Unit>> markDone(
