+++ lib/domain/workout/workout_failure.dart
