+++ lib/domain/workout/workout_failure.freezed.dart
