+++ lib/infrastructure/exercise/datasources/hive_exercise_datasource.dart
@@ -3,9 +3,9 @@ import 'package:hive/hive.dart';
 import 'package:meta/meta.dart';
 
 import 'package:power_progress/core/util/util_functions.dart';
-import 'package:power_progress/domain/core/entities/value_objects/month.dart';
-import 'package:power_progress/domain/core/entities/value_objects/one_rm.dart';
-import 'package:power_progress/domain/core/entities/week_enum.dart';
+import 'package:power_progress/domain/core/value_objects/month.dart';
+import 'package:power_progress/domain/core/value_objects/one_rm.dart';
+import 'package:power_progress/domain/core/week_enum.dart';
 import 'package:power_progress/infrastructure/exercise/models/exercise_model.dart';
 import 'package:power_progress/infrastructure/exercise/datasources/i_exercise_datasource.dart';
