+++ lib/infrastructure/one_rm/datasources/i_one_rm_datasource.dart
@@ -1,5 +1,5 @@
 import 'package:dartz/dartz.dart';
-import 'package:power_progress/domain/core/entities/value_objects/month.dart';
+import 'package:power_progress/domain/core/value_objects/month.dart';
 import 'package:power_progress/infrastructure/one_rm/models/one_rm_model.dart';
 
 abstract class IOneRmDatasource {
