+++ lib/infrastructure/workout/models/workout_done_model.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/foundation.dart';
 import 'package:dartz/dartz.dart';
 import 'package:hive/hive.dart';
-import 'package:power_progress/domain/core/entities/value_objects/month.dart';
-import 'package:power_progress/domain/core/entities/week_enum.dart';
 
-import 'package:power_progress/domain/workout/entities/workout_done.dart';
-import 'package:power_progress/domain/workout/entities/workout_failure.dart';
+import 'package:power_progress/domain/core/value_objects/month.dart';
+import 'package:power_progress/domain/core/week_enum.dart';
+import 'package:power_progress/domain/workout/workout_done.dart';
+import 'package:power_progress/domain/workout/workout_failure.dart';
 
 part 'workout_done_model.g.dart';
